+++ tests/src/WaffleRuntimeTest.php
@@ -10,8 +10,8 @@ use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Waffle\Commons\Contracts\Core\KernelInterface;
 use Waffle\Commons\Contracts\Core\TerminableInterface;
+use Waffle\Commons\Contracts\Http\GlobalsFactoryInterface;
 use Waffle\Commons\Contracts\Http\ResponseEmitterInterface;
-use Waffle\Commons\Http\Factory\GlobalsFactory;
 use Waffle\Commons\Runtime\WaffleRuntime;
 
 /**
@@ -30,7 +30,7 @@ class WaffleRuntimeTest extends AbstractTestCase
     /** @var ResponseEmitterInterface&MockObject */
     private $emitter;
 
-    /** @var GlobalsFactory&MockObject */
+    /** @var GlobalsFactoryInterface&MockObject */
     private $globalsFactory;
 
     protected function setUp(): void
@@ -41,7 +41,7 @@ class WaffleRuntimeTest extends AbstractTestCase
         $this->kernel = $this->createMock(KernelInterface::class);
         $this->request = $this->createStub(ServerRequestInterface::class);
         $this->emitter = $this->createMock(ResponseEmitterInterface::class);
-        $this->globalsFactory = $this->createMock(GlobalsFactory::class);
+        $this->globalsFactory = $this->createMock(GlobalsFactoryInterface::class);
 
         // Instantiate the runtime with mocked dependencies
         $this->runtime = new WaffleRuntime($this->globalsFactory, $this->emitter);